requirements.txt
streamlit==1.35.0
langchain==0.1.16
langchain-community==0.0.34
faiss-cpu==1.8.0
fastembed>=0.2.0
onnxruntime>=1.16.0
transformers==4.40.1
torch==2.6.0
accelerate==0.29.3
numpy==1.26.4
tqdm==4.66.4
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0